pyproject.toml
[project]
name = ""
version = "0.0.1"
dependencies = [
    "fastapi",
    "uvicorn",
    "litellm",
    "python-dotenv",
    "httpx",
    "rich",
    "pandas",
    "arize-phoenix",
    "openai",
    "openinference-instrumentation-openai",
]

[tool.uv]
package = false
